Project clean

+++ src/main/java/com/telluur/SlapBot/commands/abstractions/AbstractCommand.java
@@ -1,13 +1,10 @@
-package com.telluur.SlapBot.commands;
+package com.telluur.SlapBot.commands.abstractions;
 
 import com.jagrosh.jdautilities.command.Command;
 import com.jagrosh.jdautilities.command.CommandEvent;
 import com.jagrosh.jdautilities.commons.utils.FinderUtil;
 import com.telluur.SlapBot.SlapBot;
-import net.dv8tion.jda.core.entities.Guild;
-import net.dv8tion.jda.core.entities.Member;
-import net.dv8tion.jda.core.entities.Role;
-import net.dv8tion.jda.core.entities.User;
+import net.dv8tion.jda.core.entities.*;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -33,30 +30,31 @@ public abstract class AbstractCommand extends Command {
     @Override
     protected void execute(CommandEvent event) {
         /*
-        Scan and replace mentions in the command.
+        Scan and replace mentions in the command when used in guild.
          */
-        event.getMessage().delete().queue();
-        String filteredCmd = Arrays.stream(event.getMessage().getContentRaw().split("\\s+"))
-                .map(stringPart -> {
-                    if (stringPart.startsWith("<")) {
-                        List<Member> possibleMembers = FinderUtil.findMembers(stringPart, slapBot.getGuild());
-                        List<Role> possibleRoles = FinderUtil.findRoles(stringPart, slapBot.getGuild());
-                        if (possibleMembers.size() == 1) {
-                            Member m = possibleMembers.get(0);
-                            return String.format("`%s`", m.getEffectiveName());
-                        } else if (possibleRoles.size() == 1) {
-                            System.out.println("roles");
-                            return String.format("`%s`", possibleRoles.get(0).getName());
+        if (event.getChannel().getType().equals(ChannelType.TEXT)) {
+            event.getMessage().delete().queue();
+            String filteredCmd = Arrays.stream(event.getMessage().getContentRaw().split("\\s+"))
+                    .map(stringPart -> {
+                        if (stringPart.startsWith("<")) {
+                            List<Member> possibleMembers = FinderUtil.findMembers(stringPart, slapBot.getGuild());
+                            List<Role> possibleRoles = FinderUtil.findRoles(stringPart, slapBot.getGuild());
+                            if (possibleMembers.size() == 1) {
+                                Member m = possibleMembers.get(0);
+                                return String.format("`%s`", m.getEffectiveName());
+                            } else if (possibleRoles.size() == 1) {
+                                System.out.println("roles");
+                                return String.format("`%s`", possibleRoles.get(0).getName());
+                            } else {
+                                return stringPart;
+                            }
                         } else {
                             return stringPart;
                         }
-                    } else{
-                        return stringPart;
-                    }
-                })
-                .collect(Collectors.joining(" "));
-        event.getChannel().sendMessage(String.format("`%s` issued command: %s", event.getMember().getEffectiveName(), filteredCmd)).queue();
-
+                    })
+                    .collect(Collectors.joining(" "));
+            event.getChannel().sendMessage(String.format("`%s` issued command: %s", event.getMember().getEffectiveName(), filteredCmd)).queue();
+        }
     }
 
     boolean inValidGuildOrPrivate(CommandEvent event) {
